[SPOT-281][REFACTOR] auth를 member도메인과 분리
