+++ src/main/java/com/example/spot/auth/application/legacy/AuthService.java
@@ -1,14 +1,14 @@
-package com.example.spot.member.application.legacy.auth;
+package com.example.spot.auth.application.legacy;
 
 import com.example.spot.member.presentation.dto.MemberRequestDTO.SignUpDetailDTO;
 import com.example.spot.member.presentation.dto.rsa.Rsa;
 import com.example.spot.member.presentation.dto.MemberRequestDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.SocialLoginSignInDTO;
-import com.example.spot.member.presentation.dto.naver.NaverCallback;
-import com.example.spot.member.presentation.dto.naver.NaverOAuthToken;
-import com.example.spot.member.presentation.dto.token.TokenResponseDTO;
-import com.example.spot.member.presentation.dto.token.TokenResponseDTO.TokenDTO;
+import com.example.spot.auth.presentation.dto.naver.NaverCallback;
+import com.example.spot.auth.presentation.dto.naver.NaverOAuthToken;
+import com.example.spot.auth.presentation.dto.token.TokenResponseDTO;
+import com.example.spot.auth.presentation.dto.token.TokenResponseDTO.TokenDTO;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
