+++ src/main/java/com/example/spot/auth/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.example.spot.member.domain.auth;
+package com.example.spot.auth.domain;
 
 import com.example.spot.common.entity.BaseEntity;
 import jakarta.persistence.Entity;
