+++ src/main/java/com/example/spot/auth/domain/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.example.spot.member.domain.auth;
+package com.example.spot.auth.domain;
 
 import java.util.List;
 import java.util.Optional;
