+++ src/main/java/com/example/spot/auth/domain/TempUserDetails.java
@@ -1,24 +1,18 @@
-package com.example.spot.member.domain.auth;
+package com.example.spot.auth.domain;
 
-import java.util.Collection;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
 
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CustomUserDetails implements UserDetails {
+public class TempUserDetails implements UserDetails {
 
-    private String email; // 사용자 email
-    private Long memberId; // 사용자 이름
-    private String password; // 비밀번호
-    private boolean enabled; // 계정 활성화 여부
+    private String email; // 사용자 이름
     private Collection<? extends GrantedAuthority> authorities; // 사용자 권한
 
     @Override
@@ -28,12 +22,12 @@ public class CustomUserDetails implements UserDetails {
 
     @Override
     public String getPassword() {
-        return password;
+        return null;
     }
 
     @Override
     public String getUsername() {
-        return memberId.toString();
+        return email;
     }
 
     @Override
