+++ src/main/java/com/example/spot/auth/presentation/dto/kakao/KaKaoUser.java
@@ -1,4 +1,4 @@
-package com.example.spot.member.presentation.dto.kakao;
+package com.example.spot.auth.presentation.dto.kakao;
 
 import com.example.spot.member.domain.Member;
 import com.example.spot.member.domain.enums.Carrier;
