+++ src/main/java/com/example/spot/common/security/filters/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.example.spot.common.security.filters;
 
 import com.example.spot.common.api.exception.GeneralException;
-import com.example.spot.member.domain.auth.TempUserDetails;
+import com.example.spot.auth.domain.TempUserDetails;
 import com.example.spot.member.application.legacy.MemberService;
 import com.example.spot.common.security.utils.JwtTokenProvider;
 import com.example.spot.member.application.legacy.UserDetailsServiceCustom;
