+++ src/main/java/com/example/spot/member/application/legacy/MemberServiceImpl.java
@@ -15,19 +15,19 @@ import com.example.spot.member.domain.Member;
 import com.example.spot.member.domain.MemberRepository;
 import com.example.spot.member.presentation.dto.MemberRequestDTO;
 import com.example.spot.member.presentation.dto.MemberRequestDTO.MemberReasonDTO;
-import com.example.spot.member.domain.auth.CustomUserDetails;
-import com.example.spot.member.domain.auth.RefreshToken;
-import com.example.spot.member.domain.auth.RefreshTokenRepository;
-import com.example.spot.member.application.legacy.auth.KaKaoOAuthService;
+import com.example.spot.auth.domain.CustomUserDetails;
+import com.example.spot.auth.domain.RefreshToken;
+import com.example.spot.auth.domain.RefreshTokenRepository;
+import com.example.spot.auth.application.legacy.KaKaoOAuthService;
 import com.example.spot.member.presentation.dto.MemberResponseDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.MemberRegionDTO.RegionDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.MemberSignInDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.MemberStudyReasonDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.MemberTestDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.SocialLoginSignInDTO;
-import com.example.spot.member.presentation.dto.kakao.KaKaoOAuthToken.KaKaoOAuthTokenDTO;
-import com.example.spot.member.presentation.dto.kakao.KaKaoUser;
-import com.example.spot.member.presentation.dto.token.TokenResponseDTO.TokenDTO;
+import com.example.spot.auth.presentation.dto.kakao.KaKaoOAuthToken.KaKaoOAuthTokenDTO;
+import com.example.spot.auth.presentation.dto.kakao.KaKaoUser;
+import com.example.spot.auth.presentation.dto.token.TokenResponseDTO.TokenDTO;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
